Update dev dependencies for and builds of custom JS code
First, in dev dependency dsp-custom.js, update the header markup that is inserted on news pages, and fix the issue where the header is omitted when posts are being viewd according to category or tag. Next, update the submodule holding universal dev dependencies used to build custom JS code in every DAESA website development project. Finally, rebuild the files containing custom JS code that is ready to be deployed on the website associated with this project.