Rename: CSS Styl5t Ed3r code for WSUWP Spine theme
Rename CSS Stylesheet Editor code file «CSS/dsp-custom.prev.css» to «CSS/wsuwp-spine_css-styl5t-ed3r.prod5n.css» in anticipation of upgrading to the WDS theme for WordPress. Use this new file name pattern to include the name of the theme for which the production code was designed for.