Rename: Custom JS Editor → Prod5n code
Rename Custom JS Editor production code file «JS/dsp-build.prev.js» to «JS/wsuwp-spine_custom-js-ed3r.prod5n.js» in anticipation of upgrading to the WDS theme for WSU WordPress. Use this new file name pattern to include the name of the Spine theme for which the production code was designed for.